requirements.txt
fastapi
uvicorn
pdf2image
Pillow
python-dotenv
google-genai
python-multipart